+++ pitscout/src/components/main-page/robot-list/RobotList.scss
@@ -1,0 +1,5 @@
+
+.robot-list {
+	display: flex;
+	flex-direction: column;
+}
